backend/requirements.txt
# Core API
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
starlette==0.50.0

# Pydantic & validation
pydantic==2.12.5
pydantic_core==2.41.5
email-validator==2.3.0

# Database
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11
python-dotenv==1.2.1

# Auth
PyJWT==2.10.1
passlib==1.7.4
bcrypt<=3.2.2

# ML / Inference
torch==2.9.1
torchvision==0.24.1
diffusers==0.36.0
transformers==4.57.3
controlnet_aux==0.0.10
accelerate==1.12.0
huggingface-hub==0.36.0
safetensors==0.7.0
Pillow==12.1.0
opencv-python-headless==4.10.0.84

# Queue (optional)
celery==5.6.2
redis==7.2.1

# Utilities
requests==2.32.5
numpy==2.4.1

# Dev / Testing
pytest==9.0.2